Sql file uploded
Data analytics with Sql

Welcome to the world of data analytics with SQL! In this folder, you will find a collection of project and resources that showcase the power and versatility of SQL in data analysis.

SQL, or Structured Query Language, is a powerful tool for managing and analyzing data. It allows you to perform complex queries and extract valuable insights from large and complex data sets. In this repository, you will find examples of SQL queries and scripts that demonstrate my proficiency in data analysis using SQL.

I hope you find this repository to be informative and engaging, and I welcome any opportunity to discuss my qualifications and experience further. Thank you for visiting my repository and happy data analyzing!